Add sidebar to job card

+++ src/components/JobCard/JobBudget.jsx
@@ -1,7 +1,7 @@
 import { Label, Icon } from 'semantic-ui-react'
 
 const JobBudget = ({ budget }) => {
-  if (!budget || parseInt(budget) === 0) return null
+  if (!budget || parseFloat(budget) === 0) return null
 
   return (
     <Label>
